--- supabase/functions/HelloWorld/index.ts
@@ -1,21 +1,0 @@
-import { serve } from "https://deno.land/std@0.131.0/http/server.ts"
-import { createClient } from 'https://esm.sh/@supabase/supabase-js@1.34.0'
-
-export const supabaseClient = createClient(
-  Deno.env.get('SUPABASE_URL')!,
-  Deno.env.get('SUPABASE_ANON_KEY')!
-)
-
-serve(async (req) => {
-  const { data, error } = await supabaseClient.from('User').select()
-
-  return new Response(
-    JSON.stringify({items: data}),
-    { headers: { "Content-Type": "application/json" } },
-  )
-})
-// To invoke:
-// curl -i --location --request POST 'http://localhost:54321/functions/v1/' \
-//   --header 'Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZS1kZW1vIiwicm9sZSI6ImFub24iLCJleHAiOjE5ODM4MTI5OTZ9.CRXP1A7WOeoJeXxjNni43kdQwgnWNReilDMblYTn_I0' \
-//   --header 'Content-Type: application/json' \
-//   --data '{"name":"Functions"}'
